refactor: move user examples into bin folder under backend crate
