+++ Cargo.toml
@@ -3,6 +3,7 @@ name = "chat-app"
 version = "0.1.0"
 authors = ["Peter Satoshi <dev@ngsl.xyz>"]
 edition = "2021"
+default-run = "chat-app"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -13,3 +14,7 @@ dotenv = "0.15.0"
 
 [dependencies.backend]
 path = "backend"
+
+[[bin]]
+name = "users"
+path = "backend/src/bin/users/main.rs"
